PlayerStatusCardコンポーネントからselectedPlayerプロパティを削除し、Mainコンポーネントでの選手管理機能を改善。MatchControlPanelコンポーネントを追加し、試合管理機能を統合。
